null user protection

+++ backend/src/chat/channels/channels.controller.ts
@@ -3,6 +3,7 @@ import {
   Controller,
   Delete,
   Get,
+  InternalServerErrorException,
   Param,
   Patch,
   Post,
@@ -20,12 +21,18 @@ export class ChannelsController {
 
   @Get('all')
   async getUserChannels(@Req() req) {
+    if (!req.userID) {
+      throw new InternalServerErrorException('Bad BadRequest');
+    }
     const channels = await this.channelsService.getUserChannels(req.userID);
     return channels;
   }
 
   @Get(':id')
   async getChannelById(@Req() req, @Param('id') channelId: string) {
+    if (!req.userID) {
+      throw new InternalServerErrorException('Bad BadRequest');
+    }
     const channel = await this.channelsService.getChannelById(
       channelId,
       req.userID,
@@ -35,6 +42,9 @@ export class ChannelsController {
 
   @Post('new')
   async createChannel(@Req() req, @Body() data: newChannelDto) {
+    if (!req.userID) {
+      throw new InternalServerErrorException('Bad BadRequest');
+    }
     return this.channelsService.createChannel(data, req.userID);
   }
 
@@ -44,11 +54,17 @@ export class ChannelsController {
     @Body() data: updateChannelDto,
     @Req() req,
   ) {
+    if (!req.userID) {
+      throw new InternalServerErrorException('Bad BadRequest');
+    }
     return this.channelsService.updateChannel(req.userID, channelId, data);
   }
 
   @Delete(':id')
   async deleteChannel(@Param('id') channelId: string, @Req() req) {
+    if (!req.userID) {
+      throw new InternalServerErrorException('Bad BadRequest');
+    }
     const result = this.channelsService.deleteChannel(channelId, req.userID);
     return result;
   }
